comfysearch/apps/api/requirements.txt
flask==3.1.0
qdrant-client==1.12.1
sentence-transformers==3.3.1
pyyaml==6.0.2
requests==2.32.3
orjson==3.10.12
aiohttp==3.11.11
gunicorn==23.0.0
